templates/security_agent/requirements.txt
streamlit
requests
fpdf
discord.py
python-dotenv
openai
